+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - match Dream change to allow automatic OpenAPI generation from `delegatedAttribute` serialization of associated models
 - fix resource controller spec generator missing date and datetime in spec ensuring model owned by another user is not updated
 - resource controller spec generator supports array attributes
+- generated resource controller spec data type `DreamRequestAttributes`, not `UpdateableProperties`
 
 ## 1.10.5